requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.6
packaging==23.2
requests==2.31.0
sseclient-py==1.8.0
urllib3==2.1.0
cairosvg==2.5.2
Pillow==9.5.0
